backend/requirements.txt
absl-py==2.4.0
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
asgiref==3.11.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
backoff==2.2.1
banks==2.4.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
black==26.3.1
blinker==1.9.0
build==1.4.0
cachetools==7.0.3
cbor2==5.9.0
certifi==2026.2.25
cffi==2.0.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.4.4
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.3
cryptography==46.0.5
curl_cffi==0.13.0
cycler==0.12.1
daphne==4.1.0
databricks-sdk==0.96.0
dataclasses-json==0.6.7
dateparser==1.3.0
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
dirtyjson==1.0.8
distro==1.9.0
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docker==7.1.0
durationpy==0.10
dynaconf==3.2.12
et_xmlfile==2.0.0
evidently==0.7.20
faiss-cpu==1.13.2
Faker==40.8.0
fastapi==0.135.1
filelock==3.25.0
filetype==1.2.0
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
fonttools==4.61.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2026.2.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.4.0
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.0.0
google-genai==1.66.0
google-generativeai==0.3.2
google-pasta==0.2.0
googleapis-common-protos==1.73.0
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
greenlet==3.3.2
griffe==2.0.0
griffecli==2.0.0
griffelib==2.0.0
grpcio==1.74.0
grpcio-status==1.62.3
gunicorn==25.1.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.3.2
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
huey==2.6.0
huggingface_hub==0.36.2
hyperlink==21.0.0
idna==3.11
importlib_metadata==8.4.0
importlib_resources==6.5.2
Incremental==24.11.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kubernetes==35.0.0
langchain==0.1.16
langchain-community==0.0.38
langchain-core==0.1.53
langchain-google-genai==0.0.11
langchain-text-splitters==0.0.2
langgraph==1.1.0
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.1.147
libclang==18.1.1
litestar==2.21.0
litestar-htmx==0.5.0
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.16
llama-index-cli==0.5.5
llama-index-core==0.14.16
llama-index-embeddings-huggingface==0.6.1
llama-index-embeddings-openai==0.5.2
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.3
llama-index-llms-huggingface==0.6.1
llama-index-llms-openai==0.6.26
llama-index-readers-file==0.5.6
llama-index-readers-llama-parse==0.5.1
llama-index-workflows==2.15.1
llama-parse==0.6.54
lxml==6.0.2
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mlflow==3.10.1
mlflow-skinny==3.10.1
mlflow-tracing==3.10.1
mmh3==5.2.1
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.20.0
multidict==6.7.1
multipart==1.3.1
multitasking==0.0.12
mypy_extensions==1.1.0
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.3
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.24.3
openai==2.26.0
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
opt_einsum==3.4.0
optree==0.19.0
orjson==3.11.7
ormsgpack==1.12.2
overrides==7.7.0
packaging==26.0
pandas==2.0.0
pathspec==1.0.4
patsy==1.0.2
peewee==4.0.0
pillow==10.2.0
pip==22.3
platformdirs==4.9.2
plotly==5.24.1
polyfactory==3.3.0
posthog==7.9.12
prettytable==3.17.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==4.25.8
psutil==7.2.2
psycopg2-binary==2.9.9
pulsar-client==3.10.0
py-ubjson==0.16.1
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.13.1
Pygments==2.19.2
PyJWT==2.11.0
pyOpenSSL==26.0.0
pyparsing==3.3.2
pypdf==6.8.0
PyPika==0.51.1
pyproject_hooks==1.2.0
python-binance==1.0.17
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytokens==0.4.1
pytz==2025.2
pywin32==311
PyYAML==6.0.3
redis==7.4.0
referencing==0.37.0
regex==2026.2.28
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rich-click==1.9.7
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.4.2
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==2.5.1
service-identity==24.2.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
skops==0.13.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
sqlparse==0.5.5
starlette==0.52.1
statsmodels==0.14.0
striprtf==0.0.26
sympy==1.14.0
tenacity==8.5.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.3.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tinytag==2.2.0
tokenizers==0.19.1
torch==2.10.0
torchaudio==2.11.0
tqdm==4.67.3
transformers==4.41.2
Twisted==25.5.0
txaio==25.12.2
typer==0.24.1
typing_extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
uritemplate==4.2.0
urllib3==1.26.20
uuid_utils==0.14.1
uuid6==2025.0.1
uvicorn==0.41.0
vaderSentiment==3.3.2
waitress==3.0.2
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.6.0
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.6
wheel==0.46.3
whitenoise==6.12.0
wrapt==1.17.3
xxhash==3.6.0
yarl==1.23.0
yfinance==1.2.0
zipp==3.23.0
zope.interface==8.2
zstandard==0.25.0
